[#214] rename(profile): ProfileImageResolver 위치 변경
