+++ src/main/java/com/dduru/gildongmu/user/dto/UserInfo.java
@@ -3,7 +3,7 @@ package com.dduru.gildongmu.user.dto;
 import com.dduru.gildongmu.profile.domain.Profile;
 import com.dduru.gildongmu.profile.domain.enums.Gender;
 import com.dduru.gildongmu.profile.dto.response.ProfileImageInfo;
-import com.dduru.gildongmu.profile.service.ProfileImageResolver;
+import com.dduru.gildongmu.profile.utils.ProfileImageResolver;
 import com.dduru.gildongmu.user.domain.User;
 
 import java.time.LocalDate;
